+++ client/packages/requisitions/src/ResponseRequisition/DetailView/IndicatorEdit/IndicatorEditPage.tsx
@@ -9,7 +9,7 @@ import {
   useTranslation,
 } from '@openmsupply-client/common';
 import { useResponse } from '../../api';
-import { PageLayout } from '../PageLayout';
+import { PageLayout } from '../../../PageLayout';
 import { ListIndicatorLines } from './ListIndicators';
 import { AppRoute } from '@openmsupply-client/config';
 import { AppBarButtons } from '../ResponseLineEdit/AppBarButtons';
