+++ client/packages/requisitions/src/ResponseRequisition/DetailView/index.ts
@@ -1,3 +1,3 @@
 export { DetailView } from './DetailView';
-export { ResponseLineEditPage } from './ResponseLineEdit';
 export { IndicatorEditPage } from './IndicatorEdit/IndicatorEditPage';
+export { ResponseLineEditPage } from './ResponseLineEdit/ResponseEditPage';
